Updating the google storage files - WIP
* Removing PONG namespace :(
